requirements.txt
black==25.9.0
click==8.3.0
filelock==3.19.1
faiss-cpu==1.8.0
flake8==7.3.0
fsspec==2025.9.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas==13.0.0.19
nvidia-cuda-cupti==13.0.48
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-runtime==13.0.48
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.3.29
nvidia-cusparse==12.6.2.49
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
pdfminer.six==20240706
platformdirs==4.5.0
pycodestyle==2.14.0
pyflakes==3.4.0
pytest==8.3.3
pytokens==0.2.0
sentence-transformers==2.7.0
sympy==1.14.0
torch==2.9.0+cu130
torchvision==0.24.0+cu130
triton==3.5.0
typing_extensions==4.15.0
